+++ package.json
@@ -25,6 +25,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.2.1",
+    "react-paginate": "^8.2.0",
     "recoil": "^0.7.7",
     "recoil-persist": "^5.1.0",
     "yup": "^1.4.0"
