+++ src/app/features/flightlog/ui/atom/deleteButton/types/deleteButtonTypes.ts
@@ -1,0 +1,7 @@
+import { FlightLogDetails } from "@/app/shared/types/flightLogDetail";
+
+
+export type DeleteButtonTypes = {
+  data: FlightLogDetails;
+  iconSize: string;
+};
