+++ src/app/features/flightlog/ui/atom/editButton/types/editButtonTypes.ts
@@ -1,0 +1,4 @@
+export type EditButtonTypes = {
+  data: any;
+  iconSize: string;
+};
