+++ src/app/features/flightlog/ui/organisms/FlightLogTable/utility/formatTime.ts
@@ -1,0 +1,6 @@
+export  const formatTime = (dateString: string) => {
+  const date = new Date(dateString);
+  return `${String(date.getHours()).padStart(2, "0")}時${String(
+    date.getMinutes()
+  ).padStart(2, "0")}分`;
+};
