+++ src/app/features/flightlog/ui/template/FlightLogTemplate.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { FlightLogTable } from "../organisms/FlightLogTable";
+import { FlightLogTable } from "../organisms/FlightLogTable/FlightLogTable";
 
 export const FlightLogTemplate = () => {
   return (
