+++ src/app/features/groups/organisms/groupTable/types/GroupTableTypes.ts
